move debug_print into prints module
