+++ src/calibre/devices/kobo/books.py
@@ -9,10 +9,10 @@ from calibre import isbytestring
 from calibre.constants import DEBUG, preferred_encoding
 from calibre.devices.usbms.books import Book as Book_
 from calibre.devices.usbms.books import CollectionsBookList
-from calibre.devices.usbms.driver import debug_print
 from calibre.ebooks.metadata import author_to_author_sort
 from calibre.ebooks.metadata.book.base import Metadata
 from calibre.ebooks.metadata.book.formatter import SafeFormat
+from calibre.prints import debug_print
 from calibre.utils.config_base import prefs
