+++ src/calibre/devices/prs505/sony_cache.py
@@ -12,9 +12,9 @@ from datetime import date
 from calibre import fsync, guess_type, isbytestring, prints
 from calibre.constants import DEBUG, preferred_encoding
 from calibre.devices.errors import DeviceError
-from calibre.devices.usbms.driver import debug_print
 from calibre.ebooks.chardet import xml_to_unicode
 from calibre.ebooks.metadata import authors_to_sort_string, authors_to_string, title_sort
+from calibre.prints import debug_print
 from polyglot.binary import from_base64_bytes
 
 '''
