+++ src/calibre/gui2/device_drivers/tabbed_device_config.py
@@ -23,7 +23,7 @@ from qt.core import (
     QWidget,
 )
 
-from calibre.devices.usbms.driver import debug_print
+from calibre.prints import debug_print
 from calibre.ebooks import BOOK_EXTENSIONS
 from calibre.gui2.device_drivers.mtp_config import FormatsConfig, TemplateConfig
